requirements.txt
Django==3.0.2
djangorestframework==3.11.0
pyTelegramBotAPI==3.6.6
